+++ src/test/java/io/nats/client/AuthTests.java
@@ -23,6 +23,7 @@ import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.time.Duration;
+import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
 
 import org.junit.Test;
@@ -33,14 +34,11 @@ import io.nats.client.ConnectionListener.Events;
 public class AuthTests {
     @Test
     public void testUserPass() throws Exception {
-        String[] customArgs = {"--user","stephen","--pass","password"};
+        String[] customArgs = { "--user", "stephen", "--pass", "password" };
         try (NatsTestServer ts = new NatsTestServer(customArgs, false)) {
             // See config file for user/pass
-            Options options = new Options.Builder().
-                        server(ts.getURI()).
-                        maxReconnects(0).
-                        userInfo("stephen".toCharArray(), "password".toCharArray()).
-                        build();
+            Options options = new Options.Builder().server(ts.getURI()).maxReconnects(0)
+                    .userInfo("stephen".toCharArray(), "password".toCharArray()).build();
             Connection nc = Nats.connect(options);
             try {
                 assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
@@ -57,16 +55,12 @@ public class AuthTests {
         int port = NatsTestServer.nextPort();
         Connection nc = null;
         Subscription sub = null;
-        String[] customArgs = {"--user","stephen","--pass","password"};
+        String[] customArgs = { "--user", "stephen", "--pass", "password" };
 
         try (NatsTestServer ts = new NatsTestServer(customArgs, port, false)) {
             // See config file for user/pass
-            Options options = new Options.Builder().
-                        server(ts.getURI()).
-                        maxReconnects(-1).
-                        userInfo("stephen".toCharArray(), "password".toCharArray()).
-                        connectionListener(handler).
-                        build();
+            Options options = new Options.Builder().server(ts.getURI()).maxReconnects(-1)
+                    .userInfo("stephen".toCharArray(), "password".toCharArray()).connectionListener(handler).build();
             nc = Nats.connect(options);
             assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
 
@@ -84,10 +78,9 @@ public class AuthTests {
         }
 
         handler.waitForStatusChange(5, TimeUnit.SECONDS);
-        assertTrue("Reconnecting status", Connection.Status.RECONNECTING == nc.getStatus() ||
-                                            Connection.Status.DISCONNECTED == nc.getStatus());
+        assertTrue("Reconnecting status",
+                Connection.Status.RECONNECTING == nc.getStatus() || Connection.Status.DISCONNECTED == nc.getStatus());
         handler.prepForStatusChange(Events.RESUBSCRIBED);
-
 
         try (NatsTestServer ts = new NatsTestServer(customArgs, port, false)) {
             handler.waitForStatusChange(5, TimeUnit.SECONDS);
@@ -106,18 +99,15 @@ public class AuthTests {
     @Test
     public void testUserBCryptPass() throws Exception {
         /*
-        go run mkpasswd.go -p
-        password: password
-        bcrypt hash: $2a$11$1oJy/wZYNTxr9jNwMNwS3eUGhBpHT3On8CL9o7ey89mpgo88VG6ba
-        */
-        String[] customArgs = {"--user","ginger","--pass","$2a$11$1oJy/wZYNTxr9jNwMNwS3eUGhBpHT3On8CL9o7ey89mpgo88VG6ba"};
+         * go run mkpasswd.go -p password: password bcrypt hash:
+         * $2a$11$1oJy/wZYNTxr9jNwMNwS3eUGhBpHT3On8CL9o7ey89mpgo88VG6ba
+         */
+        String[] customArgs = { "--user", "ginger", "--pass",
+                "$2a$11$1oJy/wZYNTxr9jNwMNwS3eUGhBpHT3On8CL9o7ey89mpgo88VG6ba" };
         try (NatsTestServer ts = new NatsTestServer(customArgs, false)) {
             // See config file for user/pass
-            Options options = new Options.Builder().
-                        server(ts.getURI()).
-                        maxReconnects(0).
-                        userInfo("ginger".toCharArray(), "password".toCharArray()).
-                        build();
+            Options options = new Options.Builder().server(ts.getURI()).maxReconnects(0)
+                    .userInfo("ginger".toCharArray(), "password".toCharArray()).build();
             Connection nc = Nats.connect(options);
             try {
                 assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
@@ -130,13 +120,11 @@ public class AuthTests {
 
     @Test
     public void testUserPassInURL() throws Exception {
-        String[] customArgs = {"--user","stephen","--pass","password"};
+        String[] customArgs = { "--user", "stephen", "--pass", "password" };
         try (NatsTestServer ts = new NatsTestServer(customArgs, false)) {
             // See config file for user/pass
-            Options options = new Options.Builder().
-                        server("nats://stephen:password@localhost:"+ts.getPort()).
-                        maxReconnects(0).
-                        build();
+            Options options = new Options.Builder().server("nats://stephen:password@localhost:" + ts.getPort())
+                    .maxReconnects(0).build();
             Connection nc = Nats.connect(options);
             try {
                 assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
@@ -153,15 +141,12 @@ public class AuthTests {
         int port = NatsTestServer.nextPort();
         Connection nc = null;
         Subscription sub = null;
-        String[] customArgs = {"--user","stephen","--pass","password"};
+        String[] customArgs = { "--user", "stephen", "--pass", "password" };
 
         try (NatsTestServer ts = new NatsTestServer(customArgs, port, false)) {
             // See config file for user/pass
-            Options options = new Options.Builder().
-                        server("nats://stephen:password@localhost:"+ts.getPort()).
-                        maxReconnects(-1).
-                        connectionListener(handler).
-                        build();
+            Options options = new Options.Builder().server("nats://stephen:password@localhost:" + ts.getPort())
+                    .maxReconnects(-1).connectionListener(handler).build();
             nc = Nats.connect(options);
             assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
 
@@ -180,10 +165,9 @@ public class AuthTests {
 
         handler.waitForStatusChange(5, TimeUnit.SECONDS);
         Status status = nc.getStatus();
-        assertTrue("Reconnecting status", Connection.Status.RECONNECTING == status ||
-                                            Connection.Status.DISCONNECTED == status);
+        assertTrue("Reconnecting status",
+                Connection.Status.RECONNECTING == status || Connection.Status.DISCONNECTED == status);
         handler.prepForStatusChange(Events.RESUBSCRIBED);
-
 
         try (NatsTestServer ts = new NatsTestServer(customArgs, port, false)) {
             handler.waitForStatusChange(5, TimeUnit.SECONDS);
@@ -201,136 +185,112 @@ public class AuthTests {
 
     @Test
     public void testUserPassInURLClusteredWithDifferentUser() throws Exception {
-        String[] customArgs1 = {"--user","stephen","--pass","password"};
-        String[] customArgs2 = {"--user","alberto","--pass","casadecampo"};
+        String[] customArgs1 = { "--user", "stephen", "--pass", "password" };
+        String[] customArgs2 = { "--user", "alberto", "--pass", "casadecampo" };
         TestHandler handler = new TestHandler();
         try (NatsTestServer ts1 = new NatsTestServer(customArgs1, false);
                 NatsTestServer ts2 = new NatsTestServer(customArgs2, false)) {
             // See config file for user/pass
-            Options options = new Options.Builder().
-                        server("nats://stephen:password@localhost:"+ts1.getPort()).
-                        server("nats://alberto:casadecampo@localhost:"+ts2.getPort()).
-                        maxReconnects(4).
-                        noRandomize().
-                        connectionListener(handler).
-                        pingInterval(Duration.ofMillis(100)).
-                        build();
+            Options options = new Options.Builder().server("nats://stephen:password@localhost:" + ts1.getPort())
+                    .server("nats://alberto:casadecampo@localhost:" + ts2.getPort()).maxReconnects(4).noRandomize()
+                    .connectionListener(handler).pingInterval(Duration.ofMillis(100)).build();
             Connection nc = Nats.connect(options);
             assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
-            assertEquals(nc.getConnectedUrl(), "nats://stephen:password@localhost:"+ts1.getPort());
+            assertEquals(nc.getConnectedUrl(), "nats://stephen:password@localhost:" + ts1.getPort());
 
             handler.prepForStatusChange(Events.RESUBSCRIBED);
             ts1.close();
             handler.waitForStatusChange(2, TimeUnit.SECONDS);
 
             assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
-            assertEquals(nc.getConnectedUrl(), "nats://alberto:casadecampo@localhost:"+ts2.getPort());
+            assertEquals(nc.getConnectedUrl(), "nats://alberto:casadecampo@localhost:" + ts2.getPort());
             nc.close();
         }
     }
 
     @Test
     public void testUserPassInURLWithFallback() throws Exception {
-        String[] customArgs1 = {"--user","stephen","--pass","password"};
-        String[] customArgs2 = {"--user","alberto","--pass","casadecampo"};
+        String[] customArgs1 = { "--user", "stephen", "--pass", "password" };
+        String[] customArgs2 = { "--user", "alberto", "--pass", "casadecampo" };
         TestHandler handler = new TestHandler();
         try (NatsTestServer ts1 = new NatsTestServer(customArgs1, false);
                 NatsTestServer ts2 = new NatsTestServer(customArgs2, false)) {
             // See config file for user/pass
-            Options options = new Options.Builder().
-                        server("nats://stephen:password@localhost:"+ts1.getPort()).
-                        server("nats://localhost:"+ts2.getPort()).
-                        userInfo("alberto".toCharArray(), "casadecampo".toCharArray()).
-                        maxReconnects(4).
-                        noRandomize().
-                        connectionListener(handler).
-                        pingInterval(Duration.ofMillis(100)).
-                        build();
+            Options options = new Options.Builder().server("nats://stephen:password@localhost:" + ts1.getPort())
+                    .server("nats://localhost:" + ts2.getPort())
+                    .userInfo("alberto".toCharArray(), "casadecampo".toCharArray()).maxReconnects(4).noRandomize()
+                    .connectionListener(handler).pingInterval(Duration.ofMillis(100)).build();
             Connection nc = Nats.connect(options);
             assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
-            assertEquals(nc.getConnectedUrl(), "nats://stephen:password@localhost:"+ts1.getPort());
+            assertEquals(nc.getConnectedUrl(), "nats://stephen:password@localhost:" + ts1.getPort());
 
             handler.prepForStatusChange(Events.RESUBSCRIBED);
             ts1.close();
             handler.waitForStatusChange(2, TimeUnit.SECONDS);
 
             assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
-            assertEquals(nc.getConnectedUrl(), "nats://localhost:"+ts2.getPort());
+            assertEquals(nc.getConnectedUrl(), "nats://localhost:" + ts2.getPort());
             nc.close();
         }
     }
 
     @Test
     public void testTokenInURLClusteredWithDifferentUser() throws Exception {
-        String[] customArgs1 = {"--auth","token_one"};
-        String[] customArgs2 = {"--auth","token_two"};
+        String[] customArgs1 = { "--auth", "token_one" };
+        String[] customArgs2 = { "--auth", "token_two" };
         TestHandler handler = new TestHandler();
         try (NatsTestServer ts1 = new NatsTestServer(customArgs1, false);
                 NatsTestServer ts2 = new NatsTestServer(customArgs2, false)) {
             // See config file for user/pass
-            Options options = new Options.Builder().
-                        server("nats://token_one@localhost:"+ts1.getPort()).
-                        server("nats://token_two@localhost:"+ts2.getPort()).
-                        maxReconnects(4).
-                        noRandomize().
-                        connectionListener(handler).
-                        pingInterval(Duration.ofMillis(100)).
-                        build();
+            Options options = new Options.Builder().server("nats://token_one@localhost:" + ts1.getPort())
+                    .server("nats://token_two@localhost:" + ts2.getPort()).maxReconnects(4).noRandomize()
+                    .connectionListener(handler).pingInterval(Duration.ofMillis(100)).build();
             Connection nc = Nats.connect(options);
             assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
-            assertEquals(nc.getConnectedUrl(), "nats://token_one@localhost:"+ts1.getPort());
+            assertEquals(nc.getConnectedUrl(), "nats://token_one@localhost:" + ts1.getPort());
 
             handler.prepForStatusChange(Events.RESUBSCRIBED);
             ts1.close();
             handler.waitForStatusChange(2, TimeUnit.SECONDS);
 
             assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
-            assertEquals(nc.getConnectedUrl(), "nats://token_two@localhost:"+ts2.getPort());
+            assertEquals(nc.getConnectedUrl(), "nats://token_two@localhost:" + ts2.getPort());
             nc.close();
         }
     }
 
     @Test
     public void testTokenInURLWithFallback() throws Exception {
-        String[] customArgs1 = {"--auth","token_one"};
-        String[] customArgs2 = {"--auth","token_two"};
+        String[] customArgs1 = { "--auth", "token_one" };
+        String[] customArgs2 = { "--auth", "token_two" };
         TestHandler handler = new TestHandler();
         try (NatsTestServer ts1 = new NatsTestServer(customArgs1, false);
                 NatsTestServer ts2 = new NatsTestServer(customArgs2, false)) {
             // See config file for user/pass
-            Options options = new Options.Builder().
-                        server("nats://token_one@localhost:"+ts1.getPort()).
-                        server("nats://localhost:"+ts2.getPort()).
-                        token("token_two".toCharArray()).
-                        maxReconnects(4).
-                        noRandomize().
-                        connectionListener(handler).
-                        pingInterval(Duration.ofMillis(100)).
-                        build();
+            Options options = new Options.Builder().server("nats://token_one@localhost:" + ts1.getPort())
+                    .server("nats://localhost:" + ts2.getPort()).token("token_two".toCharArray()).maxReconnects(4)
+                    .noRandomize().connectionListener(handler).pingInterval(Duration.ofMillis(100)).build();
             Connection nc = Nats.connect(options);
             assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
-            assertEquals(nc.getConnectedUrl(), "nats://token_one@localhost:"+ts1.getPort());
+            assertEquals(nc.getConnectedUrl(), "nats://token_one@localhost:" + ts1.getPort());
 
             handler.prepForStatusChange(Events.RESUBSCRIBED);
             ts1.close();
             handler.waitForStatusChange(2, TimeUnit.SECONDS);
 
             assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
-            assertEquals(nc.getConnectedUrl(), "nats://localhost:"+ts2.getPort());
+            assertEquals(nc.getConnectedUrl(), "nats://localhost:" + ts2.getPort());
             nc.close();
         }
     }
 
     @Test
     public void testToken() throws Exception {
-        String[] customArgs = {"--auth","derek"};
+        String[] customArgs = { "--auth", "derek" };
         try (NatsTestServer ts = new NatsTestServer(customArgs, false)) {
             // See config file for user/pass
-            Options options = new Options.Builder().
-                        server(ts.getURI()).
-                        maxReconnects(0).
-                        token("derek".toCharArray()).
-                        build();
+            Options options = new Options.Builder().server(ts.getURI()).maxReconnects(0).token("derek".toCharArray())
+                    .build();
             Connection nc = Nats.connect(options);
             try {
                 assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
@@ -343,13 +303,11 @@ public class AuthTests {
 
     @Test
     public void testTokenInURL() throws Exception {
-        String[] customArgs = {"--auth","alberto"};
+        String[] customArgs = { "--auth", "alberto" };
         try (NatsTestServer ts = new NatsTestServer(customArgs, false)) {
             // See config file for user/pass
-            Options options = new Options.Builder().
-                        server("nats://alberto@localhost:"+ts.getPort()).
-                        maxReconnects(0).
-                        build();
+            Options options = new Options.Builder().server("nats://alberto@localhost:" + ts.getPort()).maxReconnects(0)
+                    .build();
             Connection nc = Nats.connect(options);
             try {
                 assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
@@ -360,17 +318,14 @@ public class AuthTests {
         }
     }
 
-    @Test(expected=AuthenticationException.class)
+    @Test(expected = AuthenticationException.class)
     public void testBadUserBadPass() throws Exception {
         Connection nc = null;
-        String[] customArgs = {"--user","stephen","--pass","password"};
+        String[] customArgs = { "--user", "stephen", "--pass", "password" };
         try (NatsTestServer ts = new NatsTestServer(customArgs, false)) {
             // See config file for user/pass
-            Options options = new Options.Builder().
-                        server(ts.getURI()).
-                        maxReconnects(0).
-                        userInfo("sam".toCharArray(), "notthepassword".toCharArray()).
-                        build();
+            Options options = new Options.Builder().server(ts.getURI()).maxReconnects(0)
+                    .userInfo("sam".toCharArray(), "notthepassword".toCharArray()).build();
             try {
                 nc = Nats.connect(options);
             } finally {
@@ -382,16 +337,13 @@ public class AuthTests {
         }
     }
 
-    @Test(expected=AuthenticationException.class)
+    @Test(expected = AuthenticationException.class)
     public void testMissingUserPass() throws Exception {
         Connection nc = null;
-        String[] customArgs = {"--user","wally","--pass","password"};
+        String[] customArgs = { "--user", "wally", "--pass", "password" };
         try (NatsTestServer ts = new NatsTestServer(customArgs, false)) {
             // See config file for user/pass
-            Options options = new Options.Builder().
-                        server(ts.getURI()).
-                        maxReconnects(0).
-                        build();
+            Options options = new Options.Builder().server(ts.getURI()).maxReconnects(0).build();
             try {
                 nc = Nats.connect(options);
             } finally {
@@ -403,17 +355,14 @@ public class AuthTests {
         }
     }
 
-    @Test(expected=AuthenticationException.class)
+    @Test(expected = AuthenticationException.class)
     public void testBadToken() throws Exception {
         Connection nc = null;
-        String[] customArgs = {"--auth","colin"};
+        String[] customArgs = { "--auth", "colin" };
         try (NatsTestServer ts = new NatsTestServer(customArgs, false)) {
             // See config file for user/pass
-            Options options = new Options.Builder().
-                        server(ts.getURI()).
-                        maxReconnects(0).
-                        token("notthetoken".toCharArray()).
-                        build();
+            Options options = new Options.Builder().server(ts.getURI()).maxReconnects(0)
+                    .token("notthetoken".toCharArray()).build();
             try {
                 nc = Nats.connect(options);
             } finally {
@@ -425,16 +374,13 @@ public class AuthTests {
         }
     }
 
-    @Test(expected=AuthenticationException.class)
+    @Test(expected = AuthenticationException.class)
     public void testMissingToken() throws Exception {
         Connection nc = null;
-        String[] customArgs = {"--auth","ivan"};
+        String[] customArgs = { "--auth", "ivan" };
         try (NatsTestServer ts = new NatsTestServer(customArgs, false)) {
             // See config file for user/pass
-            Options options = new Options.Builder().
-                        server(ts.getURI()).
-                        maxReconnects(0).
-                        build();
+            Options options = new Options.Builder().server(ts.getURI()).maxReconnects(0).build();
             try {
                 nc = Nats.connect(options);
             } finally {
@@ -453,14 +399,19 @@ public class AuthTests {
         writer.write("port: 8222"); // will get rewritten
         writer.newLine();
 
-        writer.write("authorization {"); writer.newLine();
-        writer.write("users = ["); writer.newLine();
-        writer.write(String.format("{\nnkey:%s\n}", new String(nkey))); writer.newLine();
-        writer.write("]"); writer.newLine();
-        writer.write("}"); writer.newLine();
+        writer.write("authorization {");
+        writer.newLine();
+        writer.write("users = [");
+        writer.newLine();
+        writer.write(String.format("{\nnkey:%s\n}", new String(nkey)));
+        writer.newLine();
+        writer.write("]");
+        writer.newLine();
+        writer.write("}");
+        writer.newLine();
 
         writer.close();
-        
+
         return tmp.getAbsolutePath();
     }
 
@@ -478,11 +429,8 @@ public class AuthTests {
         }
 
         try (NatsTestServer ts = new NatsTestServer(configFile, false)) {
-            Options options = new Options.Builder().
-                        server(ts.getURI()).
-                        maxReconnects(0).
-                        authHandler(new TestAuthHandler(theKey)).
-                        build();
+            Options options = new Options.Builder().server(ts.getURI()).maxReconnects(0)
+                    .authHandler(new TestAuthHandler(theKey)).build();
             Connection nc = Nats.connect(options);
             try {
                 assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
@@ -507,11 +455,8 @@ public class AuthTests {
         }
 
         try (NatsTestServer ts = new NatsTestServer(configFile, false)) {
-            Options options = new Options.Builder().
-                        server(ts.getURI()).
-                        maxReconnects(0).
-                        authHandler(Nats.staticCredentials(null, theKey.getSeed())).
-                        build();
+            Options options = new Options.Builder().server(ts.getURI()).maxReconnects(0)
+                    .authHandler(Nats.staticCredentials(null, theKey.getSeed())).build();
             Connection nc = Nats.connect(options);
             try {
                 assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
@@ -532,11 +477,8 @@ public class AuthTests {
         }
 
         try (NatsTestServer ts = new NatsTestServer("src/test/resources/operator.conf", false)) {
-            Options options = new Options.Builder().
-                        server(ts.getURI()).
-                        maxReconnects(0).
-                        authHandler(Nats.credentials("src/test/resources/jwt_nkey/user.creds")).
-                        build();
+            Options options = new Options.Builder().server(ts.getURI()).maxReconnects(0)
+                    .authHandler(Nats.credentials("src/test/resources/jwt_nkey/user.creds")).build();
             Connection nc = Nats.connect(options);
             try {
                 assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
@@ -561,11 +503,8 @@ public class AuthTests {
         String nkey = "SUAFYHVVQVOIDOOQ4MTOCTLGNZCJ5PZ4HPV5WAPROGTEIOF672D3R7GBY4";
 
         try (NatsTestServer ts = new NatsTestServer("src/test/resources/operator.conf", false)) {
-            Options options = new Options.Builder().
-                        server(ts.getURI()).
-                        maxReconnects(0).
-                        authHandler(Nats.staticCredentials(jwt.toCharArray(), nkey.toCharArray())).
-                        build();
+            Options options = new Options.Builder().server(ts.getURI()).maxReconnects(0)
+                    .authHandler(Nats.staticCredentials(jwt.toCharArray(), nkey.toCharArray())).build();
             Connection nc = Nats.connect(options);
             try {
                 assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
@@ -576,7 +515,7 @@ public class AuthTests {
         }
     }
 
-    @Test(expected=IOException.class)
+    @Test(expected = IOException.class)
     public void testBadAuthHandler() throws Exception {
         NKey theKey = NKey.createUser(null);
         assertNotNull(theKey);
@@ -590,14 +529,191 @@ public class AuthTests {
         }
 
         try (NatsTestServer ts = new NatsTestServer(configFile, false)) {
-            Options options = new Options.Builder().
-                        server(ts.getURI()).
-                        maxReconnects(0).
-                        authHandler(new TestAuthHandler(null)). // No nkey
-                        build();
+            Options options = new Options.Builder().server(ts.getURI()).maxReconnects(0)
+                    .authHandler(new TestAuthHandler(null)). // No nkey
+                    build();
             Connection nc = Nats.connect(options);
             try {
                 assertFalse("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
+            } finally {
+                nc.close();
+                assertTrue("Closed Status", Connection.Status.CLOSED == nc.getStatus());
+            }
+        }
+    }
+
+    @Test
+    public void testReconnectWithAuth() throws Exception {
+        TestHandler handler = new TestHandler();
+        String version = NatsTestServer.generateNatsServerVersionString();
+
+        if (!version.contains("v2")) {
+            // Server version doesn't support this test
+            return;
+        }
+
+        // Connect should fail on ts1
+        try (NatsTestServer ts = new NatsTestServer("src/test/resources/operator.conf", false);
+                NatsTestServer ts2 = new NatsTestServer("src/test/resources/operator.conf", false)) {
+            Options options = new Options.Builder().servers(new String[] { ts.getURI(), ts2.getURI() })
+                    .maxReconnects(-1).authHandler(Nats.credentials("src/test/resources/jwt_nkey/user.creds")).build();
+            Connection nc = Nats.connect(options);
+            try {
+                assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
+                assertEquals(ts.getURI(), nc.getConnectedUrl());
+
+                handler.prepForStatusChange(Events.RECONNECTED);
+
+                ts.close();
+
+                // Reconnect will fail because ts has the same auth error
+                handler.waitForStatusChange(5, TimeUnit.SECONDS);
+                assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
+                assertEquals(ts2.getURI(), nc.getConnectedUrl());
+            } finally {
+                nc.close();
+                assertTrue("Closed Status", Connection.Status.CLOSED == nc.getStatus());
+            }
+        }
+    }
+
+    @Test
+    public void testCloseOnReconnectWithSameError() throws Exception {
+        TestHandler handler = new TestHandler();
+        String version = NatsTestServer.generateNatsServerVersionString();
+
+        if (!version.contains("v2")) {
+            // Server version doesn't support this test
+            return;
+        }
+
+        // Connect should fail on ts1
+        try (NatsTestServer ts = new NatsTestServer("src/test/resources/operator_noacct.conf", false);
+                NatsTestServer ts2 = new NatsTestServer("src/test/resources/operator.conf", false)) {
+            Options options = new Options.Builder().servers(new String[] { ts.getURI(), ts2.getURI() })
+                    .maxReconnects(-1).connectionTimeout(Duration.ofSeconds(2))
+                    .authHandler(Nats.credentials("src/test/resources/jwt_nkey/user.creds")).build();
+            Connection nc = Nats.connect(options);
+            try {
+                assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
+                assertEquals(ts2.getURI(), nc.getConnectedUrl());
+
+                handler.prepForStatusChange(Events.CLOSED);
+
+                ts2.close();
+
+                // Reconnect will fail because ts has the same auth error
+                handler.waitForStatusChange(6, TimeUnit.SECONDS);
+                assertTrue("Closed Status", Connection.Status.CLOSED == nc.getStatus());
+            } finally {
+                nc.close();
+                assertTrue("Closed Status", Connection.Status.CLOSED == nc.getStatus());
+            }
+        }
+    }
+
+    @Test
+    public void testThatAuthErrorIsCleared() throws Exception {
+        TestHandler handler = new TestHandler();
+        String version = NatsTestServer.generateNatsServerVersionString();
+
+        if (!version.contains("v2")) {
+            // Server version doesn't support this test
+            return;
+        }
+
+        // Connect should fail on ts1
+        try (NatsTestServer ts = new NatsTestServer("src/test/resources/operator_noacct.conf", false);
+                NatsTestServer ts2 = new NatsTestServer("src/test/resources/operator.conf", false)) {
+            Options options = new Options.Builder().servers(new String[] { ts.getURI(), ts2.getURI() })
+                    .maxReconnects(-1).connectionTimeout(Duration.ofSeconds(2)).reconnectWait(Duration.ofSeconds(1)) // wait a tad to allow restarts
+                    .authHandler(Nats.credentials("src/test/resources/jwt_nkey/user.creds")).build();
+            Connection nc = Nats.connect(options);
+            try {
+                assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
+                assertEquals(ts2.getURI(), nc.getConnectedUrl());
+
+                String tsURI = ts.getURI();
+                int port = ts.getPort();
+                int port2 = ts2.getPort();
+
+                ts.close();
+
+                // ts3 will be at the same port that ts was
+                try (NatsTestServer ts3 = new NatsTestServer("src/test/resources/operator.conf", port, false)) {
+                    handler.prepForStatusChange(Events.RECONNECTED);
+
+                    ts2.close();
+
+                    // reconnect should work because we are now running with the good config
+                    handler.waitForStatusChange(5, TimeUnit.SECONDS);
+
+                    assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
+                    assertEquals(ts3.getURI(), nc.getConnectedUrl());
+                    assertEquals(tsURI, ts3.getURI());
+
+                    // Close this and go back to the bad config on that port, should be ok 1x
+                    handler.prepForStatusChange(Events.RECONNECTED);
+                    ts3.close();
+
+                    try (NatsTestServer ts4 = new NatsTestServer("src/test/resources/operator_noacct.conf", port, false);
+                            NatsTestServer ts5 = new NatsTestServer("src/test/resources/operator.conf", port2, false)) {
+                        handler.waitForStatusChange(5, TimeUnit.SECONDS);
+                        assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
+                        assertEquals(ts5.getURI(), nc.getConnectedUrl());
+                    }
+                }
+            } finally {
+                nc.close();
+                assertTrue("Closed Status", Connection.Status.CLOSED == nc.getStatus());
+            }
+        }
+    }
+
+    @Test
+    public void testReconnectAfterExpiration() throws Exception {
+        TestHandler handler = new TestHandler();
+        String version = NatsTestServer.generateNatsServerVersionString();
+
+        if (!version.contains("v2")) {
+            // Server version doesn't support this test
+            return;
+        }
+
+        CompletableFuture<Boolean> f = new CompletableFuture<Boolean>();
+
+        NatsServerProtocolMock.Customizer timeoutCustomizer = (ts, r,w) -> {
+            f.join(); // wait until we are ready
+            w.write("-ERR user authentication expired\r\n"); // Drop the line feed
+            w.flush();
+        };
+
+        int port = NatsTestServer.nextPort();
+
+        // Connect should fail on ts1
+        try (NatsServerProtocolMock ts = new NatsServerProtocolMock(timeoutCustomizer, port, true);
+                NatsTestServer ts2 = new NatsTestServer("src/test/resources/operator.conf", false)) {
+            Options options = new Options.Builder().
+                        servers(new String[] {ts.getURI(), ts2.getURI()}).
+                        maxReconnects(-1).
+                        authHandler(Nats.credentials("src/test/resources/jwt_nkey/user.creds")).
+                        build();
+            Connection nc = Nats.connect(options);
+            try {
+                assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
+                assertEquals(ts.getURI(), nc.getConnectedUrl());
+
+                handler.prepForStatusChange(Events.RECONNECTED);
+                
+                f.complete(true);
+
+                handler.waitForStatusChange(5, TimeUnit.SECONDS);
+                assertTrue("Connected Status", Connection.Status.CONNECTED == nc.getStatus());
+                assertEquals(ts2.getURI(), nc.getConnectedUrl());
+
+                String err = nc.getLastError();
+                assertNotNull(err);
+                assertTrue(err.toLowerCase().startsWith("user authentication"));
             } finally {
                 nc.close();
                 assertTrue("Closed Status", Connection.Status.CLOSED == nc.getStatus());